+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.1] - 2024-12-26
+## [1.1.2] - 2025-01-30
+
+### Fixed
+
+- Add display for required building.
+- Switch shop dialogues over to tokenized text (e.g. [LocalizedText Key]).
+- Shop displaying when no animal is purchasable
+
+## [1.1.1] - 2025-01-29
 
 ### Fixed